+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fixed a regression in the source branch dropdrown of the Start work page
 - Fixed sometimes the 'Assigned Jira work items' and 'Custom JQL filters' panels don't retrieve recently edited items
 - Fixed a bug when slash is missing after custom branch prefix in branch naming
+- Fixed bug when Jira issue view in active tab doesn't refresh itself after VS Code get focus
 
 ## What's new in 3.8.5